binance-monitor/requirements.txt
Flask
python-binance
requests
pandas
numpy
python-dotenv
